Extract hui-energy-solar-graph-card series generation into a pure module
