+++ src/panels/lovelace/cards/energy/hui-energy-solar-graph-card.ts
@@ -6,10 +6,8 @@ import { customElement, property, state } from "lit/decorators";
 import { classMap } from "lit/directives/class-map";
 import memoizeOne from "memoize-one";
 import type { BarSeriesOption, LineSeriesOption } from "echarts/charts";
-import { getEnergyColor } from "./common/color";
 import { formatNumber } from "../../../../common/number/format_number";
 import "../../../../components/chart/ha-chart-base";
-import { computeYAxisFractionDigits } from "../../../../components/chart/y-axis-fraction-digits";
 import "../../../../components/ha-card";
 import type {
   EnergyData,
@@ -19,24 +17,16 @@ import type {
 import {
   getEnergyDataCollection,
   getEnergySolarForecasts,
-  getSuggestedPeriod,
   validateEnergyCollectionKey,
 } from "../../../../data/energy";
-import type { Statistics, StatisticsMetaData } from "../../../../data/recorder";
-import { getStatisticLabel } from "../../../../data/recorder";
 import type { FrontendLocaleData } from "../../../../data/translation";
 import { SubscribeMixin } from "../../../../mixins/subscribe-mixin";
 import type { HomeAssistant } from "../../../../types";
 import type { LovelaceCard } from "../../types";
 import type { EnergySolarGraphCardConfig } from "../types";
 import { hasConfigChanged } from "../../common/has-changed";
-import {
-  computeStatMidpoint,
-  type EnergyDataPoint,
-  fillDataGapsAndRoundCaps,
-  getCommonOptions,
-  getCompareTransform,
-} from "./common/energy-chart-options";
+import { getCommonOptions } from "./common/energy-chart-options";
+import { generateEnergySolarGraphData } from "./energy-solar-graph-data";
 import type { HaECOption } from "../../../../resources/echarts/echarts";
 import "./common/hui-energy-graph-chip";
 import "../../../../components/ha-tooltip";
@@ -191,12 +181,6 @@ export class HuiEnergySolarGraphCard
   );
 
   private async _getStatistics(energyData: EnergyData): Promise<void> {
-    this._start = energyData.start;
-    this._end = energyData.end || endOfToday();
-
-    this._compareStart = energyData.startCompare;
-    this._compareEnd = energyData.endCompare;
-
     const solarSources: SolarSourceTypeEnergyPreference[] =
       energyData.prefs.energy_sources.filter(
         (source) => source.type === "solar"
@@ -213,282 +197,21 @@ export class HuiEnergySolarGraphCard
       }
     }
 
-    const datasets: (BarSeriesOption | LineSeriesOption)[] = [];
-
-    const computedStyles = getComputedStyle(this);
-
-    let yMin = Infinity;
-    let yMax = -Infinity;
-    const trackY = (v: number) => {
-      if (v < yMin) yMin = v;
-      if (v > yMax) yMax = v;
-    };
-
-    if (energyData.statsCompare) {
-      datasets.push(
-        ...this._processDataSet(
-          energyData.statsCompare,
-          energyData.statsMetadata,
-          solarSources,
-          computedStyles,
-          trackY,
-          true
-        )
-      );
-    } else {
-      // add empty dataset so compare bars are first
-      // `stack: solar` so it doesn't take up space yet
-      const firstId = solarSources[0]?.stat_energy_from ?? "placeholder";
-      datasets.push({
-        id: "compare-" + firstId,
-        type: "bar",
-        stack: "solar",
-        data: [],
-      });
-    }
-
-    datasets.push(
-      ...this._processDataSet(
-        energyData.stats,
-        energyData.statsMetadata,
-        solarSources,
-        computedStyles,
-        trackY
-      )
-    );
-
-    fillDataGapsAndRoundCaps(datasets as BarSeriesOption[]);
-
-    if (forecasts) {
-      datasets.push(
-        ...this._processForecast(
-          energyData.statsMetadata,
-          forecasts,
-          solarSources,
-          computedStyles.getPropertyValue("--primary-text-color"),
-          energyData.start,
-          energyData.end,
-          trackY
-        )
-      );
-    }
-
-    this._yAxisFractionDigits = computeYAxisFractionDigits(yMin, yMax);
-    this._chartData = datasets;
-    this._total = this._processTotal(energyData.stats, solarSources);
-  }
-
-  private _processTotal(
-    statistics: Statistics,
-    solarSources: SolarSourceTypeEnergyPreference[]
-  ) {
-    return solarSources.reduce(
-      (sum, source) =>
-        sum +
-        (source.stat_energy_from in statistics
-          ? statistics[source.stat_energy_from].reduce(
-              (acc, curr) => acc + (curr.change || 0),
-              0
-            )
-          : 0),
-      0
-    );
-  }
-
-  private _processDataSet(
-    statistics: Statistics,
-    statisticsMetaData: Record<string, StatisticsMetaData>,
-    solarSources: SolarSourceTypeEnergyPreference[],
-    computedStyles: CSSStyleDeclaration,
-    trackY: (v: number) => void,
-    compare = false
-  ) {
-    const data: BarSeriesOption[] = [];
-    const compareTransform = getCompareTransform(
-      this._start,
-      this._compareStart!
-    );
-    const period = getSuggestedPeriod(this._start, this._end);
-
-    solarSources.forEach((source, idx) => {
-      let prevStart: number | null = null;
-
-      const solarProductionData: BarSeriesOption["data"] = [];
-
-      // Process solar production data.
-      if (source.stat_energy_from in statistics) {
-        const stats = statistics[source.stat_energy_from];
-
-        for (const point of stats) {
-          if (
-            point.change === null ||
-            point.change === undefined ||
-            point.change === 0
-          ) {
-            continue;
-          }
-          if (prevStart === point.start) {
-            continue;
-          }
-          const dataPoint: EnergyDataPoint = [
-            computeStatMidpoint(
-              point.start,
-              point.end,
-              period,
-              compare ? compareTransform : undefined
-            ),
-            point.change,
-            point.start,
-          ];
-          solarProductionData.push(dataPoint);
-          trackY(point.change);
-          prevStart = point.start;
-        }
-      }
-
-      data.push({
-        type: "bar",
-        cursor: "default",
-        id: compare
-          ? "compare-" + source.stat_energy_from
-          : source.stat_energy_from,
-        name: this.hass.localize(
-          "ui.panel.lovelace.cards.energy.energy_solar_graph.production",
-          {
-            name:
-              source.name ||
-              getStatisticLabel(
-                this.hass,
-                source.stat_energy_from,
-                statisticsMetaData[source.stat_energy_from]
-              ),
-          }
-        ),
-        barMaxWidth: 50,
-        itemStyle: {
-          borderColor: getEnergyColor(
-            computedStyles,
-            this.hass.themes.darkMode,
-            false,
-            compare,
-            "--energy-solar-color",
-            idx
-          ),
-        },
-        color: getEnergyColor(
-          computedStyles,
-          this.hass.themes.darkMode,
-          true,
-          compare,
-          "--energy-solar-color",
-          idx
-        ),
-        data: solarProductionData,
-        stack: compare ? "compare" : "solar",
-      });
+    const result = generateEnergySolarGraphData({
+      hass: this.hass,
+      energyData,
+      forecasts,
+      computedStyles: getComputedStyle(this),
+      now: endOfToday(),
     });
 
-    return data;
-  }
-
-  private _processForecast(
-    statisticsMetaData: Record<string, StatisticsMetaData>,
-    forecasts: EnergySolarForecasts,
-    solarSources: SolarSourceTypeEnergyPreference[],
-    borderColor: string,
-    start: Date,
-    end: Date | undefined,
-    trackY: (v: number) => void
-  ) {
-    const data: LineSeriesOption[] = [];
-
-    const period = getSuggestedPeriod(start, end);
-
-    // Process solar forecast data.
-    solarSources.forEach((source) => {
-      if (source.config_entry_solar_forecast) {
-        const forecastsData: Record<string, number> | undefined = {};
-        source.config_entry_solar_forecast.forEach((configEntryId) => {
-          if (!forecasts![configEntryId]) {
-            return;
-          }
-          Object.entries(forecasts![configEntryId].wh_hours).forEach(
-            ([date, value]) => {
-              const dateObj = new Date(date);
-              if (dateObj < start || (end && dateObj > end)) {
-                return;
-              }
-              if (period === "month") {
-                dateObj.setDate(1);
-              }
-              if (period === "month" || period === "day") {
-                dateObj.setHours(0, 0, 0, 0);
-              } else {
-                dateObj.setMinutes(0, 0, 0);
-              }
-              const time = dateObj.getTime();
-              if (time in forecastsData) {
-                forecastsData[time] += value;
-              } else {
-                forecastsData[time] = value;
-              }
-            }
-          );
-        });
-
-        if (forecastsData) {
-          const solarForecastData: LineSeriesOption["data"] = [];
-          // Only center forecast points for sub-daily periods to align with bars.
-          // Only start timestamps available, so estimate midpoint from the gap
-          // between the first two entries. Assumes uniform spacing.
-          let forecastOffset = 0;
-          if (period === "hour" || period === "5minute") {
-            const forecastTimes = Object.keys(forecastsData)
-              .map(Number)
-              .sort((a, b) => a - b);
-            forecastOffset =
-              forecastTimes.length >= 2
-                ? (forecastTimes[1] - forecastTimes[0]) / 2
-                : 0;
-          }
-          for (const [time, value] of Object.entries(forecastsData)) {
-            const kWh = value / 1000;
-            solarForecastData.push([Number(time) + forecastOffset, kWh]);
-            trackY(kWh);
-          }
-
-          if (solarForecastData.length) {
-            data.push({
-              id: "forecast-" + source.stat_energy_from,
-              type: "line",
-              stack: "forecast",
-              name: this.hass.localize(
-                "ui.panel.lovelace.cards.energy.energy_solar_graph.forecast",
-                {
-                  name:
-                    source.name ||
-                    getStatisticLabel(
-                      this.hass,
-                      source.stat_energy_from,
-                      statisticsMetaData[source.stat_energy_from]
-                    ),
-                }
-              ),
-              step: false,
-              color: borderColor,
-              lineStyle: {
-                type: [7, 5],
-                width: 1.5,
-              },
-              symbol: "none",
-              data: solarForecastData,
-            });
-          }
-        }
-      }
-    });
-
-    return data;
+    this._start = result.start;
+    this._end = result.end;
+    this._compareStart = result.compareStart;
+    this._compareEnd = result.compareEnd;
+    this._yAxisFractionDigits = result.yAxisFractionDigits;
+    this._chartData = result.chartData;
+    this._total = result.total;
   }
 
   static styles = css`
